requirements.txt
# Core FastAPI
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Supabase (stable versions)
supabase>=2.0.0,<2.4.0
postgrest>=0.12.0,<0.14.0
httpx>=0.23.0,<0.25.0
gotrue>=2.0.0,<2.5.0

# LLM Providers
openai>=1.12.0,<2.0.0
anthropic>=0.18.0,<0.19.0
google-generativeai>=0.3.0,<0.4.0

# LangChain (for vector store)
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain-community>=0.0.20,<0.1.0

# HTTP Client
aiohttp>=3.9.0,<4.0.0

# Data Processing
numpy>=1.26.0,<2.0.0

# Voice/Speech (Deepgram)
deepgram-sdk>=3.0.0,<4.0.0

# Utilities
python-dotenv>=1.0.0

# Authentication
pyjwt>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
